Fixes the replay file copy task
The task to copy the replay file to the node was broken in that it set dest
to /opt/ndt-http-replays. We want the file to copy into:

/opt/ndt-http-replays/banjo-2016-05-10.replay

but instead it was geting copied as a file named:

/opt/ndt-http-replays

This creates the parent directory ahead of time so that the file is copied as
expected.