Adding logging support to playbooks
The client_wrapper now supports verbose logging, so this updates the playbooks
to use this option and also collect logs for each test along with the results.

On Windows, the task returns a failing error code when I try to redirect
stderr to a file, so I still need to investigate what's happening there. For
now it's a TODO.