Configures throttling as part of perf test run
Adds commands to explicitly enable packet shaping on the meddlebox before
every run of the run.yml playbook so that tests run under a defined network
speed limit.

Also changes the archive and log name schema so that the meddlebox throttling
settings are saved as part of the output archive of tests.